lapd/test: Extend test case to test msgs having data before l2h
Since e.g. the IPA input driver leaves it's specific header in front
of msg->l2h, so that msg->l2h != msg->data.

The lapdm code does not expect this at least in
rslms_rx_rll_est_req(), so the test is extended to check for this
case.

Note that the ok file reflects the current state which is not
correct.

Sponsored-by: On-Waves ehf